[feat] 앱 진입시 관심 축제에 한경대가 추가되도록 구현
TODO 제거
