+++ feature/splash/src/main/kotlin/com/unifest/android/feature/splash/SplashScreen.kt
@@ -36,7 +36,7 @@ internal fun SplashRoute(
 
     LaunchedEffect(key1 = shouldUpdate) {
         if (shouldUpdate == false) {
-            viewModel.setRecentLikedFestival()
+            viewModel.refreshFCMToken()
         }
     }
